requirements.txt
# Core dependencies
numpy>=1.19.0
scipy>=1.5.0
matplotlib>=3.3.0
pyyaml>=5.3.0
tqdm>=4.46.0

# Chess engine
python-chess>=1.3.0

# Deep learning
torch>=1.7.0
torchvision>=0.8.0
tensorboard>=2.4.0

# Computer vision
opencv-python>=4.4.0
pillow>=8.0.0

# Automation
pyautogui>=0.9.50
pynput>=1.7.0
mss>=6.1.0

# Visualization
cairosvg>=2.5.0

# Testing
pytest>=6.0.0
pytest-cov>=2.10.0

# Logging and utilities
colorlog>=4.2.0